+++ src/pysight/lst_tools.py
@@ -1,7 +1,7 @@
 import pandas as pd
 import numpy as np
 from typing import Dict
-from numba import jit
+from numba import jit, int64, uint64
 import warnings
 
 
@@ -30,165 +30,6 @@ def hex_to_bin_dict():
             'f': '1111',
         }
     return diction
-
-
-def create_data_length_dict():
-    """
-    CURRENTLY DEPRECATED
-    :return:
-    """
-    dict_of_data_length = \
-        {
-            "0": 16,
-            "5": 32,
-            "1": 32,
-            "1a": 44,
-            "2a": 48,
-            "22": 48,
-            "32": 48,
-            "2": 48,
-            "5b": 64,
-            "Db": 64,
-            "f3": 64,
-            "43": 64,
-            "c3": 64,
-            "3": 64
-        }
-    return dict_of_data_length
-
-
-def get_range(filename: str = '') -> int:
-    """
-    Finds the "range" of the current file in the proper units
-    :return: range as defined my MCS, after bit depth multiplication
-    """
-    import re
-
-    if filename == '':
-        raise ValueError('No filename given.')
-
-    format_range = re.compile(r'range=(\d+)')
-    with open(filename, 'r') as f:
-        cur_str = f.read(500)  # read 100 chars for the basic values
-
-    range_before_bit_depth = int(re.search(format_range, cur_str).group(1))
-    format_bit_depth = re.compile(r'bitshift=(\w+)')
-    bit_depth_wrong_base = re.search(format_bit_depth, cur_str).group(1)
-    bit_depth_as_hex = bit_depth_wrong_base[-2:]  # last 2 numbers count
-    range_after_bit_depth = range_before_bit_depth * 2 ** int(bit_depth_as_hex, 16)
-
-    assert isinstance(range_after_bit_depth, int)
-    return range_after_bit_depth
-
-
-def get_timepatch(filename: str = '') -> str:
-    """
-    Get the time patch value out of of a list file.
-    :param filename: File to be read.
-    :return: Time patch value as string.
-    """
-    import re
-
-    if filename == '':
-        raise ValueError('No filename given.')
-
-    format_timepatch = re.compile(r'time_patch=(\w+)')
-    with open(filename, 'r') as f:
-        cur_str = f.read(5000)  # read 5000 chars for the timepatch value
-
-    timepatch = re.search(format_timepatch, cur_str).group(1)
-    # data_length_dict = create_data_length_dict()
-    # data_length = data_length_dict[timepatch]
-
-    assert isinstance(timepatch, str)
-    # assert isinstance(data_length, int)
-    # return timepatch, data_length - DEPRECATED
-    return timepatch
-
-
-def get_start_pos(filename: str = '') -> int:
-    """
-    Returns the start position of the data
-    :param filename: Name of file
-    :return: Integer of file position for f.seek() method
-    """
-    import re
-
-    if filename == '':
-        raise ValueError('No filename given.')
-
-    format_data = re.compile(r"DATA]\n")
-    pos_in_file = 0
-    with open(filename, 'r') as f:
-        while pos_in_file == 0:
-            line = f.readline()
-            match = re.search(format_data, line)
-            if match is not None:
-                pos_in_file = f.tell()
-                return pos_in_file  # to have the [DATA] as header
-
-
-def read_lst_file_debug(filename: str='', start_of_data_pos: int=0, num_of_lines=0) -> pd.DataFrame:
-    """
-    Read the list file into a dataframe.
-    :param filename: Name of list file.
-    :param start_of_data_pos: The place in chars in the file that the data starts at.
-    :return: Dataframe with all events registered.
-    """
-    from itertools import islice
-
-    if filename is '' or start_of_data_pos == 0:
-        return ValueError('Wrong input detected.')
-
-    with open(filename, 'r') as f:
-        f.seek(start_of_data_pos)
-        n_lines = list(islice(f, int(num_of_lines)))
-        if not n_lines:
-            pass
-        file_separated = [st.rstrip() for st in n_lines]
-    df = pd.DataFrame(file_separated, columns=['raw'], dtype=str)
-
-    assert df.shape[0] > 0
-    return df
-
-
-def read_lst_file(filename: str = '', start_of_data_pos: int = 0) -> pd.DataFrame:
-    """
-    Read the list file into a dataframe.
-    :param filename: Name of list file.
-    :param start_of_data_pos: The place in chars in the file that the data starts at.
-    :return: Dataframe with all events registered.
-    """
-    # TRIAL VERSION, DEPRECATED. TRY TO RAISE FROM DEAD ONLY IF CURRENT VERSION IS SLOW ###############
-    # def binarize(str1):
-    #     return "{0:0{1}b}".format(int(str1, 16), data_length)
-    #
-    # with open(filename, 'r') as f:
-    #     dict_bin = dict([('bin', binarize)])
-    #     f.seek(start_of_data_pos)
-    #     df = pd.read_csv(f, header=0, converters=dict_bin, dtype=str, names=['bin'])
-    ######################################################################################################
-
-    if filename is '' or start_of_data_pos == 0:
-        return ValueError('Wrong input detected.')
-
-    with open(filename, 'r') as f:
-        f.seek(start_of_data_pos)
-        file_separated = [line.rstrip() for line in f]  # TODO: SciLuigi?
-        # TODO: f = io.open(r'C:\Users\Hagai\Documents\GitHub\python-pysight\multiscalerAsPreampMicrogliaMouseFLIM035.lst', 'rb')
-        # f.seek(1590)
-        # numread = 0
-        #
-        # for i in itertools.count():
-        #     numread = f.readinto(b)
-        #     lista.append(b.decode())
-        #     if not numread:
-        #         break
-        # array.array can also work, with fromfile()
-    df = pd.DataFrame(file_separated, columns=['raw'], dtype=str)
-
-    assert df.shape[0] > 0
-    return df
 
 
 def timepatch_sort(df, timepatch: str='', data_range: int=0, input_channels: Dict=None) -> pd.DataFrame:
@@ -283,31 +124,6 @@ def create_line_array(last_event_time: int=None, num_of_lines=None, num_of_frame
     return line_array
 
 
-def create_inputs_dict(gui=None) -> Dict:
-    """
-    Create a dictionary for all input channels. Currently allows for three channels.
-    'Empty' channels will not be checked.
-    """
-
-    if gui is None:
-        raise ValueError('No GUI received.')
-
-    dict_of_inputs = {}
-
-    if gui.input_start.get() != 'Empty':
-        dict_of_inputs[gui.input_start.get()] = '110'
-
-    if gui.input_stop1.get() != 'Empty':
-        dict_of_inputs[gui.input_stop1.get()] = '001'
-
-    if gui.input_stop2.get() != 'Empty':
-        dict_of_inputs[gui.input_stop2.get()] = '010'
-
-    assert len(dict_of_inputs) >= 1
-    assert 'Empty' not in list(dict_of_inputs.keys())
-
-    return dict_of_inputs
-
 def validate_created_data_channels(dict_of_data: Dict):
     """
     Make sure that the dictionary that contains all data channels makes sense.
@@ -360,7 +176,10 @@ def determine_data_channels(df: pd.DataFrame=None, dict_of_inputs: Dict=None,
         # NUMBA SORT NOT WORKING:
         # sorted_vals = numba_sorted(relevant_values.values)
         # dict_of_data[key] = pd.DataFrame(sorted_vals, columns=['abs_time'])
-        dict_of_data[key] = relevant_values.sort_values().reset_index(drop=True)
+        if 'Tag Lens' == key:
+            dict_of_data[key] = relevant_values.sort_values().reset_index(drop=True)
+        else:
+            dict_of_data[key] = relevant_values.reset_index(drop=True)
         # TODO: GroupBy the lines above?
 
     if 'Lines' not in dict_of_data.keys():  # A 'Lines' channel has to exist to create frames
@@ -376,7 +195,8 @@ def determine_data_channels(df: pd.DataFrame=None, dict_of_inputs: Dict=None,
                                          pixels=x_pixels, spacing_between_lines=spacing_between_lines)
         dict_of_data['Frames'] = pd.Series(frame_array, name='abs_time', dtype=np.uint64)
     else:  # Add 0 to the first entry of the series
-        dict_of_data['Frames'] = pd.Series([0], name='abs_time').append(dict_of_data['Frames'], ignore_index=True)
+        dict_of_data['Frames'] = pd.Series([0], name='abs_time', dtype=np.uint64).append(dict_of_data['Frames'],
+                                                                                         ignore_index=True)
 
     # Validations
     validate_created_data_channels(dict_of_data)
@@ -384,7 +204,7 @@ def determine_data_channels(df: pd.DataFrame=None, dict_of_inputs: Dict=None,
     return dict_of_data
 
 
-@jit(nopython=True, cache=True)
+@jit((int64[:](uint64[:], uint64[:])), nopython=True, cache=True)
 def numba_search_sorted(input_sorted, input_values):
     """ Numba-powered searchsorted function. """
     return np.searchsorted(input_sorted, input_values) - 1